requirements.txt
pandas>=2.0.0
google-cloud-bigquery>=3.0.0
db-dtypes>=1.0.0
apache-airflow>=2.5.0
pandas-gbq>=0.19.0